requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
